Issue 664 Create BLAST database when reference is registered
