+++ server/ngb-cli/profiles/server.properties
@@ -39,3 +39,5 @@ formats_url=/restapi/dataitem/formats
 version=@version@
 get_existing_index_url=/restapi/getPathToExistingIndex
 project_description_url=/restapi/project/%d/description
+register_blast_database_url=/restapi/blast/database
+create_blast_database_url=/restapi/createdb
