+++ server/ngb-cli/src/main/java/com/epam/ngb/cli/manager/command/handler/http/AbstractHTTPCommandHandler.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2016 EPAM Systems
+ * Copyright (c) 2016-2021 EPAM Systems
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,8 +24,18 @@
 
 package com.epam.ngb.cli.manager.command.handler.http;
 
-import static com.epam.ngb.cli.constants.MessageConstants.*;
-import static com.epam.ngb.cli.entity.BiologicalDataItemResourceType.*;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_DATAITEM_FORMATS_NOT_FOUND;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_FAILED_TO_LOAD_USER;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_FILE_NOT_FOUND;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_INDEX_REQUIRED;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_PERMISSIONS_NOT_FOUND;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_PROJECT_NOT_FOUND;
+import static com.epam.ngb.cli.constants.MessageConstants.ERROR_REFERENCE_NOT_FOUND;
+import static com.epam.ngb.cli.constants.MessageConstants.ILLEGAL_PATH_FORMAT;
+import static com.epam.ngb.cli.constants.MessageConstants.SEVERAL_RESULTS_FOR_QUERY;
+import static com.epam.ngb.cli.constants.MessageConstants.getMessage;
+import static com.epam.ngb.cli.entity.BiologicalDataItemResourceType.FILE;
+import static com.epam.ngb.cli.entity.BiologicalDataItemResourceType.getTypeFromPath;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -38,7 +48,20 @@ import java.util.List;
 import java.util.Map;
 
 import com.epam.ngb.cli.app.Utils;
-import com.epam.ngb.cli.entity.*;
+import com.epam.ngb.cli.entity.AclClass;
+import com.epam.ngb.cli.entity.AclSecuredEntry;
+import com.epam.ngb.cli.entity.BiologicalDataItem;
+import com.epam.ngb.cli.entity.BiologicalDataItemFormat;
+import com.epam.ngb.cli.entity.IDList;
+import com.epam.ngb.cli.entity.NgbUser;
+import com.epam.ngb.cli.entity.Project;
+import com.epam.ngb.cli.entity.RequestPayload;
+import com.epam.ngb.cli.entity.ResponseResult;
+import com.epam.ngb.cli.entity.Role;
+import com.epam.ngb.cli.entity.SpeciesEntity;
+import com.epam.ngb.cli.entity.UserContext;
+import com.epam.ngb.cli.entity.blast.BlastDatabaseVO;
+import com.epam.ngb.cli.entity.blast.CreateDatabaseRequest;
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.math.NumberUtils;
@@ -824,16 +847,16 @@ public abstract class AbstractHTTPCommandHandler extends AbstractSimpleCommandHa
         return Pair.of(fileAbsolutePath, index);
     }
 
-    protected void registerBlastDatabase(final BlastDatabaseVO databaseVO) {
+    protected void createBlastDatabase(final CreateDatabaseRequest createDatabaseRequest) {
         HttpPost request = (HttpPost) getRequestFromURLByType("POST",
-                getServerParameters().getServerUrl() + getServerParameters().getRegisterBlastDatabaseUrl());
-        String result = getPostResult(databaseVO, request);
+                getServerParameters().getServerUrl() + getServerParameters().getCreateBlastDatabaseUrl());
+        String result = getPostResult(createDatabaseRequest, request);
         isResultOk(result);
     }
 
-    protected void createBlastDatabase(final BlastDatabaseVO databaseVO) {
+    protected void registerBlastDatabase(final BlastDatabaseVO databaseVO) {
         HttpPost request = (HttpPost) getRequestFromURLByType("POST",
-                getServerParameters().getServerUrl() + getServerParameters().getCreateBlastDatabaseUrl());
+                getServerParameters().getServerUrl() + getServerParameters().getRegisterBlastDatabaseUrl());
         String result = getPostResult(databaseVO, request);
         isResultOk(result);
     }
