+++ server/ngb-cli/src/main/java/com/epam/ngb/cli/manager/command/handler/http/blast/BlastDatabaseRegistrationHandler.java
@@ -28,7 +28,7 @@ package com.epam.ngb.cli.manager.command.handler.http.blast;
 
 import com.epam.ngb.cli.app.ApplicationOptions;
 import com.epam.ngb.cli.constants.MessageConstants;
-import com.epam.ngb.cli.entity.BlastDatabaseVO;
+import com.epam.ngb.cli.entity.blast.BlastDatabaseVO;
 import com.epam.ngb.cli.manager.command.handler.Command;
 import com.epam.ngb.cli.manager.command.handler.http.AbstractHTTPCommandHandler;
 import org.apache.http.client.methods.HttpPost;
@@ -65,7 +65,12 @@ public class BlastDatabaseRegistrationHandler extends AbstractHTTPCommandHandler
             throw new IllegalArgumentException(MessageConstants.getMessage(
                     ILLEGAL_DATABASE_SOURCE, source));
         }
-        database = new BlastDatabaseVO(arguments.get(0), arguments.get(1), type, source);
+        database = BlastDatabaseVO.builder()
+                .name(arguments.get(0))
+                .path(arguments.get(1))
+                .type(type)
+                .source(source)
+                .build();
     }
 
     @Override public int runCommand() {
